// src/types/spot.d.ts
export interface BasicTimeRangeParam {
    startTime?: number;
    endTime?: number;
}
export interface BasicFromPaginatedParams {
    fromId?: number;
    startTime?: number;
    endTime?: number;
    limit?: number;
}
export type SymbolStatus = 'PRE_TRADING' | 'TRADING' | 'POST_TRADING' | 'END_OF_DAY' | 'HALT' | 'AUCTION_MATCH' | 'BREAK';
export interface SystemStatusResponse {
    status: 0 | 1;
    msg: 'normal' | 'system maintenance';
}
export interface DailyAccountSnapshotParams {
    type: 'SPOT' | 'MARGIN' | 'FUTURES';
    startTime?: number;
    endTime?: number;
    limit?: number;
}
// ...
export interface DailySpotAccountSnapshot {
    data: {
        balances: SpotBalance[];
        totalAssetOfBtc: numberInString;
    };
    type: 'spot';
    updateTime: number;
}
// ...
export type DailyAccountSnapshotElement = DailySpotAccountSnapshot | DailyMarginAccountSnapshot | DailyFuturesAccountSnapshot;
export interface DailyAccountSnapshot {
    code: number;
    msg: string;
    snapshotVos: DailyAccountSnapshotElement[];
}
export interface CoinNetwork {
    addressRegex: string;
    coin: string;
    depositDesc: string;
    depositEnable: boolean;
    isDefault: boolean;
    memoRegex: string;
    minConfirm: number;
    name: string;
    network: string;
    resetAddressStatus: boolean;
    specialTips?: string;
    specialWithdrawTips?: string;
    unlockConfirm: number;
    withdrawDesc: string;
    withdrawEnable: boolean;
    withdrawFee: numberInString;
    withdrawMin: numberInString;
    withdrawMax: numberInString;
    withdrawInternalMin: numberInString;
    withdrawIntegerMultiple: numberInString;
    depositDust?: numberInString;
    sameAddress: boolean;
    estimatedArrivalTime: number;
    busy: boolean;
    contractAddressUrl?: string;
    contractAddress?: string;
}
export interface AllCoinsInformationResponse {
    coin: string;
    depositAllEnable: boolean;
    free: numberInString;
    freeze: numberInString;
    ipoable: numberInString;
    ipoing: numberInString;
    isLegalMoney: boolean;
    locked: numberInString;
    name: string;
    networkList: CoinNetwork[];
    storage: numberInString;
    trading: boolean;
    withdrawAllEnable: boolean;
    withdrawing: numberInString;
}
export interface WithdrawParams {
    coin: string;
    withdrawOrderId?: string;
    network?: string;
    address: string;
    addressTag?: string;
    amount: number;
    transactionFeeFlag?: boolean;
    name?: string;
    walletType?: number;
}
// ...
export type DepositStatusCode = `${EnumDepositStatus}`;
export interface DepositHistoryParams {
    coin?: string;
    status?: DepositStatusCode;
    startTime?: number;
    endTime?: number;
    offset?: number;
    limit?: number;
}
export interface DepositHistory {
    amount: numberInString;
    coin: string;
    network: string;
    status: number;
    address: string;
    addressTag: string;
    txId: string;
    insertTime: number;
    transferType: number;
    confirmTimes: string;
}
export declare enum EnumWithdrawStatus {
    EmailSent = 0,
    Cancelled = 1,
    AwaitingApproval = 2,
    Rejected = 3,
    Processing = 4,
    Failure = 5,
    Completed = 6
}
export type WithdrawStatusCode = `${EnumWithdrawStatus}`;
export interface WithdrawHistoryParams {
    coin?: string;
    withdrawOrderId?: string;
    status?: WithdrawStatusCode;
    offset?: number;
    limit?: number;
    startTime?: number;
    endTime?: number;
}
export declare enum EnumWithdrawTransferType {
    External = 0,
    Interal = 1
}
export type WithdrawTransferType = `${EnumWithdrawTransferType}`;
export interface WithdrawHistory {
    address: string;
    amount: numberInString;
    applyTime: string;
    coin: string;
    id: string;
    withdrawOrderId: string;
    network: string;
    transferType: WithdrawTransferType;
    status: number;
    transactionFee: numberInString;
    txId: string;
}
export interface DepositAddressParams {
    coin: string;
    network?: string;
}
export interface DepositAddressResponse {
    address: string;
    coin: string;
    tag: string;
    url: string;
}
export interface ConvertDustParams {
    asset: string[];
}
// ...
export type UniversalTransferType = `${EnumUniversalTransferType}`;
export interface UniversalTransferParams {
    type: UniversalTransferType;
    asset: string;
    amount: number;
    fromSymbol: string;
    toSymbol: string;
}
export interface UniversalTransferHistoryParams {
    type: UniversalTransferType;
    startTime?: number;
    endTime?: number;
    current?: number;
    size?: number;
}
export interface ExchangeInfoParams {
    symbol?: string;
    symbols?: string[];
    permissions?: string | string[];
    showPermissionSets?: boolean;
    symbolStatus?: string;
}
export interface NewSpotOrderParams<T extends OrderType = OrderType, RT extends OrderResponseType | undefined = OrderResponseType> {
    symbol: string;
    side: OrderSide;
    type: T;
    timeInForce?: OrderTimeInForce;
    quantity?: number;
    quoteOrderQty?: number;
    price?: number;
    newClientOrderId?: string;
    strategyId?: number;
    strategyType?: number;
    stopPrice?: number;
    trailingDelta?: number;
    icebergQty?: number;
    newOrderRespType?: RT;
    isIsolated?: StringBoolean;
    sideEffectType?: SideEffects;
}
// ...
export interface GetOCOParams {
    symbol?: string;
    isIsolated?: StringBoolean;
    orderListId?: number;
    origClientOrderId?: string;
}
export interface NewSpotSOROrderParams {
    symbol: string;
    side: OrderSide;
    type: OrderType;
    timeInForce?: OrderTimeInForce;
    quantity: number;
    price?: number;
    newClientOrderId?: string;
    strategyId?: number;
    strategyType?: number;
    icebergQty?: number;
    newOrderRespType?: OrderResponseType;
    selfTradePreventionMode?: SelfTradePreventionMode;
}
export type APILockTriggerCondition = 'GCR' | 'IFER' | 'UFR';
export interface APITriggerConditionSymbolStatus {
    i: APILockTriggerCondition;
    c: number;
    v: number;
    t: number;
}
export interface APITradingStatus {
    data: {
        isLocked: boolean;
        plannedRecoverTime: number;
        triggerCondition: Record<APILockTriggerCondition, number>;
        indicators: Record<ExchangeSymbol, APITriggerConditionSymbolStatus[]>;
        updateTime: number;
    };
}
export interface APIPermissions {
    ipRestrict: boolean;
    createTime: number;
    enableWithdrawals: boolean;
    enableInternalTransfer: boolean;
    permitsUniversalTransfer: boolean;
    enableVanillaOptions: boolean;
    enableReading: boolean;
    enableFutures: boolean;
    enableMargin: boolean;
    enableSpotAndMarginTrading: boolean;
    tradingAuthorityExpirationTime: number;
    enableFixApiTrade: boolean;
    enableFixReadOnly: boolean;
}
export interface AssetDetail {
    minWithdrawAmount: numberInString;
    depositStatus: boolean;
    withdrawFee: number;
    withdrawStatus: boolean;
    depositTip?: string;
}
export interface SymbolTradeFee {
    symbol: string;
    makerCommission: numberInString;
    takerCommission: numberInString;
}
export interface SymbolExchangeInfo {
    symbol: string;
    status: string;
    baseAsset: string;
    baseAssetPrecision: number;
    quoteAsset: string;
    quotePrecision: number;
    quoteAssetPrecision: number;
    baseCommissionPrecision: number;
    quoteCommissionPrecision: number;
    orderTypes: OrderType[];
    icebergAllowed: boolean;
    ocoAllowed: boolean;
    quoteOrderQtyMarketAllowed: boolean;
    allowTrailingStop: boolean;
    cancelReplaceAllowed: boolean;
    isSpotTradingAllowed: boolean;
    isMarginTradingAllowed: boolean;
    filters: SymbolFilter[];
    permissions: ('SPOT' | 'MARGIN')[];
    defaultSelfTradePreventionMode: SelfTradePreventionMode;
    allowedSelfTradePreventionModes: SelfTradePreventionMode[];
}
export interface ExchangeInfo {
    timezone: string;
    serverTime: number;
    rateLimits: RateLimiter[];
    exchangeFilters: ExchangeFilter[];
    symbols: SymbolExchangeInfo[];
}
export interface OrderBookResponse {
    lastUpdateId: number;
    bids: OrderBookRow[];
    asks: OrderBookRow[];
}
export interface RawTrade {
    id: number;
    price: numberInString;
    qty: numberInString;
    quoteQty: numberInString;
    time: number;
    isBuyerMaker: boolean;
    isBestMatch: boolean;
}
export interface RawAccountTrade {
    symbol: string;
    id: number;
    orderId: number;
    orderListId: number;
    price: numberInString;
    qty: numberInString;
    quoteQty: numberInString;
    commission: numberInString;
    commissionAsset: string;
    time: number;
    isBuyer: boolean;
    isMaker: boolean;
    isBestMatch: boolean;
}
export interface AggregateTrade {
    a: number;
    p: numberInString;
    q: numberInString;
    f: number;
    l: number;
    T: number;
    m: boolean;
    M: boolean;
}
export interface CurrentAvgPrice {
    mins: number;
    price: numberInString;
}
export interface DailyChangeStatistic {
    symbol: string;
    priceChange: numberInString;
    priceChangePercent: numberInString;
    weightedAvgPrice: numberInString;
    prevClosePrice: numberInString;
    lastPrice: numberInString;
    lastQty: numberInString;
    bidPrice: numberInString;
    bidQty: numberInString;
    askPrice: numberInString;
    askQty: numberInString;
    openPrice: numberInString;
    highPrice: numberInString;
    lowPrice: numberInString;
    volume: numberInString;
    quoteVolume: numberInString;
    openTime: number;
    closeTime: number;
    firstId: number;
    lastId: number;
    count: number;
}
export interface SymbolOrderBookTicker {
    symbol: string;
    bidPrice: numberInString;
    bidQty: numberInString;
    askPrice: numberInString;
    askQty: numberInString;
}
export interface OrderResponseACK {
    symbol: string;
    orderId: number;
    orderListId: number;
    clientOrderId: string;
    transactTime: number;
}
export interface OrderResponseResult {
    symbol: string;
    orderId: number;
    orderListId: number;
    clientOrderId: string;
    transactTime: number;
    price: numberInString;
    origQty: numberInString;
    executedQty: numberInString;
    cummulativeQuoteQty: numberInString;
    status: OrderStatus;
    timeInForce: OrderTimeInForce;
    type: OrderType;
    side: OrderSide;
    workingTime: number;
    selfTradePreventionMode: SelfTradePreventionMode;
}
export interface OrderFill {
    price: numberInString;
    qty: numberInString;
    commission: numberInString;
    commissionAsset: string;
}
export interface OrderResponseFull {
    symbol: string;
    orderId: number;
    orderListId?: number;
    clientOrderId: string;
    transactTime: number;
    price: numberInString;
    origQty: numberInString;
    executedQty: numberInString;
    cummulativeQuoteQty: numberInString;
    status: OrderStatus;
    timeInForce: OrderTimeInForce;
    type: OrderType;
    side: OrderSide;
    marginBuyBorrowAmount?: number;
    marginBuyBorrowAsset?: string;
    isIsolated?: boolean;
    workingTime: number;
    selfTradePreventionMode: SelfTradePreventionMode;
    fills: OrderFill[];
}
// ...
export interface CancelSpotOrderResult {
    symbol: string;
    origClientOrderId: string;
    orderId: number;
    orderListId: number;
    clientOrderId: string;
    transactTime: number;
    price: numberInString;
    origQty: numberInString;
    executedQty: numberInString;
    cummulativeQuoteQty: numberInString;
    status: OrderStatus;
    timeInForce: OrderTimeInForce;
    type: OrderType;
    side: OrderSide;
    isIsolated?: boolean;
    selfTradePreventionMode: SelfTradePreventionMode;
}
// ...
export interface SpotOrder {
    symbol: string;
    orderId: number;
    orderListId: number;
    clientOrderId: string;
    price: numberInString;
    origQty: numberInString;
    executedQty: numberInString;
    cummulativeQuoteQty: numberInString;
    status: OrderStatus;
    timeInForce: OrderTimeInForce;
    type: OrderType;
    side: OrderSide;
    stopPrice: numberInString;
    icebergQty: numberInString;
    time: number;
    updateTime: number;
    isWorking: boolean;
    origQuoteOrderQty: numberInString;
    isIsolated?: boolean;
}
export interface SpotAssetBalance {
    asset: string;
    free: numberInString;
    locked: numberInString;
}
export interface AccountInformation {
    makerCommission: number;
    takerCommission: number;
    buyerCommission: number;
    sellerCommission: number;
    canTrade: boolean;
    canWithdraw: boolean;
    canDeposit: boolean;
    updateTime: number;
    accoountType: string;
    balances: SpotAssetBalance[];
    permissions: any[];
}
// ...
export interface SubAccountListParams {
    email?: string;
    isFreeze?: StringBoolean;
    page?: number;
    limit?: number;
}